requirements.txt
torch>=2.0.0
transformers>=4.30.0
polars>=0.19.0
bitsandbytes>=0.41.0
accelerate>=0.20.0
datasets>=2.12.0
huggingface-hub>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0
jupyter>=1.0.0
ipykernel>=6.0.0
python-dotenv>=1.0.0
requests>=2.31.0
peft>=0.4.0
huggingface_hub>=0.20.3 